refactor: split ingredient ui off med request ui
